refactor(audit-docs): split into modular components (types, parsers, audit, report)
